+++ sys/slb.c
@@ -221,8 +221,8 @@ load_and_init_slb(char *name, char *path, long min_ver, SHARED_LIB **sl, int *co
 		DEBUG(("Slbopen: SLB with shared text bit?!"));
 		r = ENOEXEC;
 slb_error:
-		sys_m_free((long) b->p_env);
-		sys_m_free((long) b);
+		sys_m_free(b->p_env);
+		sys_m_free(b);
 		if (--mr->links == 0)
 			free_region(mr);
 		return(r);
@@ -291,7 +291,7 @@ slb_error:
 	if (hitpa < b->p_hitpa)
 	{
 		b->p_hitpa = hitpa;
-		r = sys_m_shrink(0, (long) b, b->p_hitpa - (long)b);
+		r = sys_m_shrink(0, b, b->p_hitpa - (long)b);
 		if (r)
 		{
 			DEBUG(("Slbopen: Couldn't shrink basepage"));
